Separate events, professors, students and subject by author

+++ flaskcalendar/professors/routes.py
@@ -5,14 +5,16 @@ from flaskcalendar.models import Professor, Subject, ProfessorSubjects
 from flask_login import login_required, current_user
 from flaskcalendar.professors.forms import AddProfessorForm
 from flaskcalendar.main.utils import addToHistory
-
+from flaskcalendar.users.utils import isAuthor
 professorsAPP = Blueprint('professors', __name__)
+
+
 
 @professorsAPP.route("/add_professor", methods=['GET','POST'])
 @login_required
 def add_professor():
     form = AddProfessorForm()
-    sList = Subject.query.filter_by()
+    sList = Subject.query.filter_by(author_id=current_user.id)
     if form.validate_on_submit():
         author_id = int(current_user.id)
         instance = Professor(name=form.name.data, last_name=form.last_name.data, email=form.email.data, phone=form.phone.data, author_id=author_id)
@@ -31,6 +33,23 @@ def add_professor():
     return render_template('professors/add_professor.html',form=form, subjectsList=sList)
 
 
+@professorsAPP.route("/professor")
+def professor():
+    if request.method == 'GET' and request.args.get('id'):
+        professor_id = int(request.args.get('id'))
+        instance = Professor.query.filter_by(id=professor_id).first()
+        if isAuthor(instance):
+            return render_template('professors/professor.html',title=f'{instance.fullName()} Events', instance=instance)
+    return redirect(url_for('professors.professors'))
+
+
+@professorsAPP.route("/professors")
+def professors():
+    professorsList = []
+    if current_user.is_authenticated:
+        professorsList = Professor.query.filter_by(author_id=current_user.id)
+    return render_template('professors/professors.html', pList = professorsList)
+
 
 @professorsAPP.route("/edit_professor", methods=['GET','POST'])
 @login_required
@@ -39,45 +58,36 @@ def edit_professor():
     if form.validate_on_submit():
         professor_id = int(request.form.get("instance_id"))
         instance = Professor.query.filter_by(id=professor_id).first()
-        instance.name = form.name.data
-        instance.last_name  = form.last_name.data
-        instance.email  = form.email.data
-        instance.phone = form.phone.data
-        addToHistory(instance,'edit')
-        db.session.commit()
-        flash(f"{instance.fullName()} account has been updated",'success')
-        return redirect(url_for('professors.professors'))
+        if isAuthor(instance):
+            instance.name = form.name.data
+            instance.last_name  = form.last_name.data
+            instance.email  = form.email.data
+            instance.phone = form.phone.data
+            addToHistory(instance,'edit')
+            db.session.commit()
+            flash(f"{instance.fullName()} account has been updated",'success')
+            return redirect(url_for('professors.professors'))
     elif request.method == 'GET' and request.args.get('Professor'):
         professor_id = int(request.args.get('Professor'))
         instance = Professor.query.filter_by(id=professor_id).first()
-        form.name.data = instance.name
-        form.last_name.data = instance.last_name
-        form.email.data = instance.email
-        form.phone.data = instance.phone
-        return render_template('professors/edit_professor.html', form=form, instance=instance)
+        if isAuthor(instance):
+            form.name.data = instance.name
+            form.last_name.data = instance.last_name
+            form.email.data = instance.email
+            form.phone.data = instance.phone
+            return render_template('professors/edit_professor.html', form=form, instance=instance)
     return redirect(url_for('professors.professors'))
-
-@professorsAPP.route("/professor")
-def professor():
-    if request.method == 'GET' and request.args.get('id'):
-        professor_id = int(request.args.get('id'))
-        instance = Professor.query.filter_by(id=professor_id).first()
-        return render_template('professors/professor.html',title=f'{instance.fullName()} Events', instance=instance)
-    return redirect(url_for('professors.professors'))
-
-
-@professorsAPP.route("/professors")
-def professors():
-    professorsList = Professor.query
-    return render_template('professors/professors.html', pList = professorsList)
 
 
 @professorsAPP.route("/professors/<int:professor_id>/delete", methods=['POST'])
 @login_required
 def professor_delete(professor_id):
     instance = Professor.query.get_or_404(professor_id)
-    addToHistory(instance,'delete')
-    db.session.delete(instance)
-    db.session.commit()
-    flash(f"Professor has been deleted correctly",'success')
+    if isAuthor(instance):
+        addToHistory(instance,'delete')
+        db.session.delete(instance)
+        db.session.commit()
+        flash(f"Professor has been deleted correctly",'success')
+    else:
+        flash(f"You're not the author of this Professor",'warning')
     return redirect(url_for('professors.professors'))
